// src/app/api/generated/models/tour-log-weather-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CoordinateDto } from '../models/coordinate-dto';
export interface TourLogWeatherDto {
  fetchedAt?: string;
  lookupCoordinate?: CoordinateDto;
  precipitationMm?: number;
  provider?: string;
  providerDataset?: string;
  relativeHumidityPercent?: number;
  temperatureC?: number;
  tourLogId?: number;
  weatherCode?: number;
  weatherDescription?: string;
  weatherObservedAt?: string;
  windSpeedKmh?: number;
}